refactor: separate files for config and home
Simplify by extracting configuration processing (command line flags,
read and process configuration file) and home page processing from
main.go, into their own files config.go and main.go. Also created
VSCode workspace file
