+++ .vscode/gowebapp.code-workspace
@@ -1,0 +1,8 @@
+{
+	"folders": [
+		{
+			"path": "/Users/peterplamondon/go/src/github.com/peterpla/gowebapp"
+		}
+	],
+	"settings": {}
+}
